address comments from code review
