+++ zebra-chain/src/lib.rs
@@ -8,6 +8,7 @@
 extern crate serde;
 
 mod merkle_tree;
+mod serde_helpers;
 mod sha256d_writer;
 
 pub mod addresses;
@@ -27,20 +28,8 @@ pub use redjubjub;
 #[cfg(test)]
 use proptest_derive::Arbitrary;
 
-use serde_helpers::*;
-
 /// The size of an Equihash solution in bytes (always 1344).
 const EQUIHASH_SOLUTION_SIZE: usize = 1344;
-
-mod serde_helpers {
-    use super::EQUIHASH_SOLUTION_SIZE;
-    use serde_big_array::big_array;
-
-    big_array! {
-        BigArray;
-        +EQUIHASH_SOLUTION_SIZE, 580, 601, 80
-    }
-}
 
 /// An enum describing the possible network choices.
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, Serialize, Deserialize)]
