+++ zebra-chain/src/proofs/groth16.rs
@@ -1,10 +1,12 @@
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
+use serde::{Deserialize, Serialize};
 use std::{fmt, io};
 
+use crate::serde_helpers;
 use crate::serialization::{SerializationError, ZcashDeserialize, ZcashSerialize};
 
 /// An encoding of a Groth16 proof, as used in Zcash.
-pub struct Groth16Proof(pub [u8; 192]);
+#[derive(Serialize, Deserialize)]
+pub struct Groth16Proof(#[serde(with = "serde_helpers::BigArray")] pub [u8; 192]);
 
 impl fmt::Debug for Groth16Proof {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -46,53 +48,6 @@ impl ZcashDeserialize for Groth16Proof {
         let mut bytes = [0; 192];
         reader.read_exact(&mut bytes[..])?;
         Ok(Self(bytes))
-    }
-}
-
-impl<'de> Deserialize<'de> for Groth16Proof {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct Groth16Visitor;
-        impl<'de> serde::de::Visitor<'de> for Groth16Visitor {
-            type Value = Groth16Proof;
-            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-                formatter.write_str("192 bytes of data")
-            }
-
-            fn visit_seq<A>(self, mut seq: A) -> Result<Groth16Proof, A::Error>
-            where
-                A: serde::de::SeqAccess<'de>,
-            {
-                let mut bytes = [0u8; 192];
-                {
-                    let bytes: &mut [_] = &mut bytes;
-                    for (i, byte) in bytes.iter_mut().enumerate() {
-                        *byte = seq.next_element()?.ok_or_else(|| {
-                            serde::de::Error::invalid_length(i, &"expected 192 bytes")
-                        })?;
-                    }
-                }
-                Ok(Groth16Proof(bytes))
-            }
-        }
-
-        deserializer.deserialize_tuple(192, Groth16Visitor)
-    }
-}
-
-impl Serialize for Groth16Proof {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        use serde::ser::SerializeTuple;
-        let mut tup = serializer.serialize_tuple(192)?;
-        for byte in &self.0[..] {
-            tup.serialize_element(byte)?;
-        }
-        tup.end()
     }
 }
