+++ zebra-chain/src/transaction/shielded_data.rs
@@ -8,6 +8,7 @@ use crate::note_commitment_tree::SaplingNoteTreeRootHash;
 use crate::notes::sapling;
 use crate::proofs::Groth16Proof;
 use crate::redjubjub::{self, Binding, SpendAuth};
+use crate::serde_helpers;
 
 /// A _Spend Description_, as described in [protocol specification §7.3][ps].
 ///
@@ -65,19 +66,6 @@ impl Arbitrary for Spend {
     type Strategy = BoxedStrategy<Self>;
 }
 
-#[derive(Deserialize, Serialize)]
-#[serde(remote = "jubjub::AffinePoint")]
-struct AffinePoint {
-    #[serde(getter = "jubjub::AffinePoint::to_bytes")]
-    bytes: [u8; 32],
-}
-
-impl From<AffinePoint> for jubjub::AffinePoint {
-    fn from(local: AffinePoint) -> Self {
-        jubjub::AffinePoint::from_bytes(local.bytes).unwrap()
-    }
-}
-
 /// A _Output Description_, as described in [protocol specification §7.4][ps].
 ///
 /// [ps]: https://zips.z.cash/protocol/protocol.pdf#outputencoding
@@ -92,7 +80,7 @@ pub struct Output {
     /// XXX refine to a specific type.
     pub cmu: [u8; 32],
     /// An encoding of an ephemeral Jubjub public key.
-    #[serde(with = "AffinePoint")]
+    #[serde(with = "serde_helpers::AffinePoint")]
     pub ephemeral_key: jubjub::AffinePoint,
     /// A ciphertext component for the encrypted output note.
     pub enc_ciphertext: sapling::EncryptedCiphertext,
@@ -135,13 +123,6 @@ impl Arbitrary for Output {
     type Strategy = BoxedStrategy<Self>;
 }
 
-#[derive(Deserialize, Serialize)]
-#[serde(remote = "Either")]
-enum LocalEither<A, B> {
-    Left(A),
-    Right(B),
-}
-
 /// Sapling-on-Groth16 spend and output descriptions.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct ShieldedData {
@@ -154,7 +135,7 @@ pub struct ShieldedData {
     /// separately, as the [`ShieldedData::spends`] and [`ShieldedData::outputs`]
     /// methods provide iterators over all of the [`SpendDescription`]s and
     /// [`Output`]s.
-    #[serde(with = "LocalEither")]
+    #[serde(with = "serde_helpers::Either")]
     pub first: Either<Spend, Output>,
     /// The rest of the [`Spend`]s for this transaction.
     ///
